distribute QueryBuilderOperatorLoader
